+++ lib/recorder/changeset.rb
@@ -33,7 +33,7 @@ module Recorder
       @previous_version = item.dup
 
       changes.each do |key, change|
-        @previous_version.send("#{key}=", change[0])
+        @previous_version.try("#{key}=", change[0]) if pair?(change)
       end
 
       @previous_version
@@ -49,10 +49,16 @@ module Recorder
       @next_version = item.dup
 
       changes.each do |key, change|
-        @next_version.send("#{key}=", change[1])
+        @next_version.try("#{key}=", change[1]) if pair?(change)
       end
 
       @next_version
+    end
+
+    private
+
+    def pair?(change)
+      change.is_a?(Array) && change.size == 2
     end
   end
 end